+++ CHANGES.rst
@@ -12,6 +12,9 @@ Changelog
   remove redundant code in ``browser/ia``.
   [boulch]
 
+- Changer Event content title  for sponsors and partners
+  [thomlamb]
+
 
 1.4.52 (2026-06-23)
 -------------------